+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 0.3
+* Add identify and is translatable options to translate node
+* Add pulse option for shine node
+* Add random color to shine color option
+* Add play sound file option to speak node
+* Add type of tone to return to analyze tone node
+
+## 0.2.1
+* Add option to flip photo when taking a photo with see node
+
 ## 0.2.0
 * Add additional motions for the wave node: armback, lowerarm, raisearm
 * Add speak node